FIX/rename component
