+++ src/components/TableVideosStaticSection.tsx
